ef refactoring
