+++ ChineseDuck.Import/EfModels/Word.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace ChineseDuck.Import
+namespace ChineseDuck.Import.EfModels
 {
     public partial class Word
     {
