+++ ChineseDuck.Import/EfModels/WordFileO.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 
-namespace ChineseDuck.Import
+namespace ChineseDuck.Import.EfModels
 {
     public partial class WordFileO
     {
