+++ ChineseDuck.Import/EfModels/WordFileP.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 
-namespace ChineseDuck.Import
+namespace ChineseDuck.Import.EfModels
 {
-    public partial class WordFileT
+    public partial class WordFileP
     {
         public long IdWord { get; set; }
         public Guid Id { get; set; }
