+++ ChineseDuck.Import/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using Microsoft.Extensions.Configuration;
 
 namespace ChineseDuck.Import
@@ -7,9 +8,16 @@ namespace ChineseDuck.Import
     {
         static void Main(string[] args)
         {
-            var builder = new ConfigurationBuilder();
+            var builder = new ConfigurationBuilder()
+                .SetBasePath(Directory.GetCurrentDirectory())
+                .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true);
+
+            IConfigurationRoot configuration = builder.Build();
+
+            Console.WriteLine(configuration.GetConnectionString("OldSqlDb"));
 
             Console.WriteLine("Hello World!");
+            Console.ReadKey();
         }
     }
 }
